# executor: reuse chunk for GetMatchedRowsAndPtrs calls (#48090)